IMP_PFM-13_BEMEJIA: Inicio micro medidor virtual

+++ backend/services/entrenamiento/app/models.py
@@ -1,5 +1,5 @@
 from sqlmodel import Field, SQLModel
-from datetime import datetime, date
+from datetime import date
 from typing import Optional
 
 class Entrenamiento(SQLModel, table=True):
@@ -8,7 +8,7 @@ class Entrenamiento(SQLModel, table=True):
     sport_type: str  # Tipo de deporte: 'atletismo' o 'ciclismo'
     duration: str
     fecha: date
-    calories_active: int = 0
-    total_calories: int = 0
+    calories_active: float = 0.0
+    total_calories: float = 0.0
     fcm: int = 0
